fix(simd.h): Address NEON issues
Primarily, recent changes (PR 4071) to vint4::store for the NEON case
appear to have some type mismatches, which apple clang on ARM-based
Mac (including our CI) seems ok with, but which is generating type
errors on other ARM Linux platforms.

I think the types were weird here, so I tightened it up to get the
types right for temporary variables in that function. That's the
primary fix here.

Secondarily, I modfied simd.h and the CMake setup so that build option
USE_SIMD=0 will disable NEON in the same way that it disables SSE. (I
realized that USE_SIMD=0 was not disabling NEON, so there was no way
for a NEON platform to completely disable SIMD if they needed to.)

Signed-off-by: Larry Gritz <lg@larrygritz.com>