build: Fixes for DCMTK (#4147)
It somehow escaped our notice, but not only does DCMTK have an exported
config, but also CMake itself has a FindDCMTK.cmake module. So get rid
of ours, and prefer the config if available. This is hoped to solve some
edge cases for finding and using this library properly.

Fixes #4146
(I hope?)

Signed-off-by: Larry Gritz <lg@larrygritz.com>