fix(simd.h): gather_mask was wrong for no-simd fallback
The operation it's supposed to be performing is to merge in the
gathered value where the mask is set, and leave mask=0 lanes
unchanged. But for our "no SIMD" fallback (i.e., no AVX2 for 4- and
8-wide), we incorrectly implemented it as setting the unmasked lanes
to 0!

This was quite wrong, and unfortunately masked by the fact that our
test of the function initialized a variable to 0. Change to a better
number. The best number.

Signed-off-by: Larry Gritz <lg@larrygritz.com>