refactor(simd.h): use default for ctrs and assignment where applicable
simd.h improvements and refactorization for consistency:

* Use `= default` for all the uninitialized constructors, copy
  constructors, and assignment from the same type.
* Fix the `operator=` to be more standard by returning a `vec&`
  instead of `const vec&`, which apparently is unusual.
* CI: build one of the Mac tests with avx2 support (the other one is
  only SSE).
* Test that these all result in C++ understanding that these classes
  are trivially copyable.

There is a bit of awkwardness where for the vint types, we have to
define the copy constructors for the Intel icc compiler. It just seems
to do the wrong thing for `=default` and fails lots of our tests.  I
don't understand why, and I'm not inclined to investigate too deeply,
since it's a discontinued compiler we support only for legacy reasons,
so it doesn't bother me too much that we special-case it. One reason I
think it's a compiler bug is that it has no problem with the floating
point vector types. It's confused about something with the simd ints.

Signed-off-by: Larry Gritz <lg@larrygritz.com>