Add Windows VS17 test case
Disable SIMD acceleration of fast_exp2 when using MSVS prior to 2022.
It seems to sporadically generate bad code in this function and I
can't seem to restructure the function in a way that consistently is
correct. Doesn't seem to happen with newer MSVS, nor on any other
platform or compiler. I'm assuming this is a compiler bug and it's not
worth pursuing for an old version of one compiler, for this one
function that isn't really used all that often or in
performance-critical paths.

Some extra slosh in the simd fast_exp test.

Rewrite OIIO_CHECK_SIMD_EQUAL_THRESH to print the differences and use
more digits of precision when it fails.

Signed-off-by: Larry Gritz <lg@larrygritz.com>