build: make an OpenImageIO_Util_static library and target
For the OpenImageIO_Util library, build a separately named static
version, even if otherwise we are building dynamic libraries.

This can be a help to applications that only need OpenImageIO_Util,
and not the whole OpenImageIO library, and furthermore would prefer a
static library so they don't need a runtime dependency on our dynamic
libraries.

Cavat emptor: Beware that ustring in particular has singletons inside!
It could be a bad thing to have an application that contains both the
static and dynamic versions of libOpenImageIO_Util, or to have
multiple copies of the static library inside. Beware particularly
about linking multiple libraries, which each might contain a static
OpenImageIO_util_static use the static util library in an application.
You have been warned!

Signed-off-by: Larry Gritz <lg@larrygritz.com>