build: switch to target-based definitions
We had lots of stray usage of add_defintions, add_compile_options,
etc.  In modern cmake, this is considered bad form, since it makes
subprojects pollute the global option lists. A better way is to add
all these to specific targets, so now that's what we do in this PR.
This should make all the definitions and options we want to set for
compmilng OIIO's code not have any interference with subprojects, or
with higher-level surrounding projecs.

Signed-off-by: Larry Gritz <lg@larrygritz.com>