build: have checked_find_package accept CONFIG, fix DCMTK
Add another optional flag to our checked_find_packge cmake function --
CONFIG means only accept a config for the package, never use a
FindBlah.cmake module.

(Reminder: this is in contrast to the PREFER_CONFIG, which tries a
config first, then if that fails, will try the module, and also
different from the default if neither is used, which is to prefer the
module over the exported config.)

And rig our search for DCMTK to use the new CONFIG option. There is a
FindDCMTK.cmake modules that comes with cmake, and for at least some
versions of cmake, it seems subtly broken in such a way that actually
wrecks our finding of other packages properly.

Signed-off-by: Larry Gritz <lg@larrygritz.com>