Fix out-of-bounds reads when using OpenEXR decreasingY lineOrder.
OpenEXR expects to process scanlines in decreasing order when the
decreasingY lineOrder option is enabled. This change fixes the code
that provides the chunks of scanlines to OpenEXR so that the proper
scanlines are available when it accesses the FrameBuffer OIIO created.
The old code was providing incorrect scanlines AND setting up the
FrameBuffer with incorrect pointers so out-of-bounds reads were
occurring.

Signed-off-by: Aaron Colwell <300262+acolwell@users.noreply.github.com>