build!: Move OpenEXR/Imath minimum to 3.1
These are build-breaking changes that impose a new minimum
OpenEXR/Imath version supported of 3.1 or higher.

Needless to say, this will not be backported to any release branches,
which guarantee that they will never raise the minimum supported
versions of any dependencies. This is just for future OIIO. I won't
merge this until after April 1, at which point I will also tag the
last spot in master that is known to work with OpenEXR 2.x.

Signed-off-by: Larry Gritz <lg@larrygritz.com>