Remove boost thread_specific_ptr (#4221)
Remove the dependency on boost's thread_specific_ptr by using hashmap's
of static duration between the owning object and the target datatype.

After digging a bit into how boost's pointer works behind the scenes, it
is actually a very similar mechanism, with the lookup being into an
std::map. I don't have an easy way to measure performance here, but none
of these are on a hot path (with the possible exception of the
per_thread_info for the image cache, but this one can be managed
externally if you want the best performance.

---------

Signed-off-by: Chris Kulla <ckulla@gmail.com>
Signed-off-by: Vlad (Kuzmin) Erium <libalias@gmail.com>