fix: Error retrieval safeguards for recycled objects
The recent switch from boost thread_specific_ptr left a bug: if an
ImageInput (say) is deleted without the caller clearing its error
state, and then later another ImageInput is allocated and happens to
get the same address as the first one, it can "inherit" the
unretrieved error state left behind by the other.

Fix by clearing the error state at construction/initialization.

This still doesn't 100% fix all possible cases. If the calling thread
that left behind an unretrieved error in the first object is a
different thread than the one that creates the next recycled object,
there is still some cruft left behind.

I feel like we still don't have a foolproof replacement for the really
nice lifetime management that thread_specific_ptr was doing for us.
One could argue that unretrieved errors in an object are a "bug" in
the user's program, but it's an exceptionally easy mistake to make and
shouldn't result in heisenbehavor that differs based on whether
allocations happen to use the same addresses for other objects.

Signed-off-by: Larry Gritz <lg@larrygritz.com>