fix: fix crash when no default fonts are found
The nullptr at the end of default_font_name was left over from a time
that we iterated by going through array elements until we hit the
nullptr that signalled we had passed the last valid element. At some
point we switched to a range-for, which iterates over every element of
the array without needing an "end sentinel", and we should have
eliminated the nullptr but failed to do so.

Fixes 4248

Signed-off-by: Larry Gritz <lg@larrygritz.com>