build: More warning elimination for clang18
This is a continuation of PR 4246, had one more spot where we needed
to disable warnings for unity builds involving clang18 and deprecated
unicode conversion functions.

There's some changes included to ci.yml and install_homebrew_deps.bash
that are part of work I did to make it easier for me to force it to
use clang-18 for certain tests while working on this.

Signed-off-by: Larry Gritz <lg@larrygritz.com>