build: Add CMath target for the sake of static libtiff
Recent libtiff versions export cmake configs that, if you built
libtiff as a static library, will reference targets it set up when it
built, but which are nonstandard names and won't exist when you build
and consume their config exports. Ugh!

Some can be swept under the rug by turning off some libtiff optional
dependencies, but I can't seem to get rid of a target CMath::CMath
that static libtiff target wants. It just points to the math library.
So I'm just setting that up so it will be found when we do the
find_package(tiff). Oh, the hoops we jump through for badly
constructed packges.

Signed-off-by: Larry Gritz <lg@larrygritz.com>